pyproject.toml
[project]
name = "ynab-mcp-server"
version = "0.1.0"
description = "FastMCP-based MCP server for YNAB covering all endpoints via OpenAPI"
authors = [{ name = "troylar" }]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

# PEP 621-compliant dependency array
dependencies = [
  "fastmcp>=2.8.0",
  "httpx>=0.25,<1.0",
  "pydantic>=2.6,<3.0",
  "PyYAML>=6.0",
]

[project.urls]
Homepage = "https://api.ynab.com/v1"
Repository = "https://github.com/troylar/ynab-mcp-server"

[project.scripts]
ynab-mcp-server = "ynab_mcp_server.main:cli"

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
  "pytest-asyncio>=0.23,<1",
  "pytest-cov>=5,<6",
  "respx>=0.21,<1",
  "anyio>=4.4,<5",
  "ruff>=0.5",
  "mypy>=1.10",
  "types-PyYAML>=6.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = []

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
strict_optional = true

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"

[tool.uv]
# If using uv, this helps ensure reproducible runs
dev-dependencies = [
  "pytest>=8,<9",
  "pytest-asyncio>=0.23,<1",
  "respx>=0.21,<1",
  "anyio>=4.4,<5",
  "ruff>=0.5",
  "mypy>=1.10",
  "types-PyYAML>=6.0",
]
